+++ src/validation.rs
@@ -113,6 +113,9 @@ pub enum MeshValidationErrorKind {
         "Vertex weights for mesh {mesh_name:?} are not normalized. Vertex weights should sum to 1.0."
     )]
     VertexWeightsNotNormalized { mesh_name: String },
+
+    #[error("Mesh {mesh_name:?} has vertex weights with a weight of 0.0 that can be removed.")]
+    VertexWeightsZero { mesh_name: String },
 }
 
 pub struct SkelValidationError;
@@ -748,11 +751,17 @@ fn validate_mesh_subindices(validation: &mut ModelValidationErrors, mesh: &MeshD
 
 fn validate_mesh_vertex_weights(validation: &mut ModelValidationErrors, mesh: &MeshData) {
     for (i, o) in mesh.objects.iter().enumerate() {
+        let mut has_zero_weights = false;
+
         // TODO: Also validate 4 weights per vertex.
         let mut weight_total_by_vertex = HashMap::<u32, f32>::new();
         for influence in &o.bone_influences {
             for w in &influence.vertex_weights {
                 *weight_total_by_vertex.entry(w.vertex_index).or_default() += w.vertex_weight;
+
+                if w.vertex_weight == 0.0 {
+                    has_zero_weights = true;
+                }
             }
         }
 
@@ -764,6 +773,16 @@ fn validate_mesh_vertex_weights(validation: &mut ModelValidationErrors, mesh: &M
             let error = MeshValidationError {
                 mesh_object_index: i,
                 kind: MeshValidationErrorKind::VertexWeightsNotNormalized {
+                    mesh_name: o.name.clone(),
+                },
+            };
+            validation.mesh_errors.push(error);
+        }
+
+        if has_zero_weights {
+            let error = MeshValidationError {
+                mesh_object_index: i,
+                kind: MeshValidationErrorKind::VertexWeightsZero {
                     mesh_name: o.name.clone(),
                 },
             };
@@ -1927,6 +1946,57 @@ Use a Source Color of "One" or use a shader that does not premultiply alpha."#,
 
         assert_eq!(
             r#"Vertex weights for mesh "a" are not normalized. Vertex weights should sum to 1.0."#,
+            format!("{}", validation.mesh_errors[0])
+        );
+    }
+
+    #[test]
+    fn mesh_vertex_weights_zero() {
+        let mesh = MeshData {
+            major_version: 1,
+            minor_version: 10,
+            objects: vec![
+                MeshObjectData {
+                    name: "a".to_owned(),
+                    subindex: 0,
+                    bone_influences: vec![BoneInfluence {
+                        bone_name: "bone".to_owned(),
+                        vertex_weights: vec![
+                            VertexWeight {
+                                vertex_index: 0,
+                                vertex_weight: 1.0,
+                            },
+                            VertexWeight {
+                                vertex_index: 0,
+                                vertex_weight: 0.0,
+                            },
+                        ],
+                    }],
+                    ..Default::default()
+                },
+                MeshObjectData {
+                    name: "b".to_owned(),
+                    subindex: 0,
+                    ..Default::default()
+                },
+            ],
+        };
+
+        let mut validation = ModelValidationErrors::default();
+        validate_mesh_vertex_weights(&mut validation, &mesh);
+
+        assert_eq!(
+            vec![MeshValidationError {
+                mesh_object_index: 0,
+                kind: MeshValidationErrorKind::VertexWeightsZero {
+                    mesh_name: "a".to_owned(),
+                }
+            }],
+            validation.mesh_errors
+        );
+
+        assert_eq!(
+            r#"Mesh "a" has vertex weights with a weight of 0.0 that can be removed."#,
             format!("{}", validation.mesh_errors[0])
         );
     }